// src/app/model/electroInterface.ts
 export  interface Electro {

    rut: string ;
    nombre: string;
    obs?: string;
    urgencia?: string;
    imagen?: string;
    linkImagen: string;
    id: string;
    edad: string;
    fecha?: any;
    fechaInforme: Date;
    fechaInImpresion: Date;
    ficha: string;
    informado: string;
    informe?: string;
}